+++ Tests/Helper/Mocks/MockKeyboardAccessoryDelegate.swift
@@ -1,0 +1,32 @@
+//
+//  MockKeyboardAccessoryDelegate.swift
+//  KeyboardSupport-iOSTests
+//
+//  Created by Earl Gaspard on 1/4/19.
+//  Copyright © 2019 Bottle Rocket. All rights reserved.
+//
+
+import Foundation
+import KeyboardSupport
+
+enum TapType {
+    case back
+    case next
+    case done
+}
+
+class MockKeyboardAccessoryDelegate: KeyboardAccessoryDelegate {
+    var tapType: TapType?
+    
+    func keyboardAccessoryDidTapBack(_ accessory: UIView) {
+        tapType = .back
+    }
+    
+    func keyboardAccessoryDidTapNext(_ accessory: UIView) {
+        tapType = .next
+    }
+    
+    func keyboardAccessoryDidTapDone(_ accessory: UIView) {
+        tapType = .done
+    }
+}